Implement Encode for primitives and collections
